moved the router components in their own folder. small cleanup. added few tests
